Backups of the original WordPress site hosted on IONOS
These include a native Wordpress data export as well as compressed backups pushed by UpdraftPlus (and the accompanying log file).

Note that when I went to install UpdraftPlus WordPress reported it could not write to the directory... turns out the `plugins` directory was missing. I recreated it and corrected the permissions so the install could complete, but could not find the original directory. It's possible one of the IONOS security flags quarantined it, but I'm not sure where and could find no record of that. I should be able to figure out if there were plugins in use from the database records though.